remove a lot of useless code for listeners
Cyclone core library guarantees that at most one listener call takes place on a given
entity at any one time.  So no point in having a mutex to prevent that in the C++ binding.

The ListenerArg object wasn't used, eliminating it means no need to worry about freeing
it.

Disabled entities don't exist in the core of Cyclone (at this moment), so a half-baked
attempt at implementing it just for listeners in the C++ binding makes no sense.

Rewriting the order of setting the listeners means there's no value in keeping the
"dds_listener_t" around. We can always request it from the core library.

We can remove the listeners before breaking down the C++ entity, that should guarantee
that the memory for the entity still exists (and I suspect we can still dereference it,
although I suspect it causes undefined behaviour ...).  There was some code to deal with
such bad timing between destructors and listeners, but I doubt it really covered all the
cases.

It still suffers from the possibility of deadlocking if a listener invocation happens to
end up holding the last strong reference to the entity: then the call to dds_set_listener
will block because it is being called from within the listener.  (And if somehow that
function could make progress, then dds_delete would block ...)